Language Convergence Infrastructure committed.
This is the heart of SLPS for now.
It uses extractors to get BGFs from various FL implementations.
It uses grammar transformations (XBGF) to approximate grammars.
It uses grammar diffing (BDT) to make sure grammars converge.
It uses the language document (LDT) as source of test set.
It uses differential testing to compare how different implementations react to that test set.
It reports results in a concise manner.
It generates a PDF diagram showing what grammars converged using what transformations.
It operates on basis of a simple configuration file.

For now it does not work since XBGF deal is not yet committed.
Future work: migrate from CFG to XML.



git-svn-id: https://slps.svn.sourceforge.net/svnroot/slps@89 ab42f6e0-554d-0410-b580-99e487e6eeb2